refactor: 重新组织事件模块结构
- 事件信号基类 `src/lib/xboyEvents/EventSignal.ts` -> `src/core/events/signal.ts`
- 玩家准备好接收广播事件（通过视角检测） `src/lib/xboyEvents/ready.ts` -> `src/features/events/player-ready.ts`
